+++ src/links.rs
@@ -72,6 +72,7 @@ pub mod lists {
     pub const SUBSCRIBE: &'static str = "https://api.twitter.com/1.1/lists/subscribers/create.json";
     pub const UNSUBSCRIBE: &'static str = "https://api.twitter.com/1.1/lists/subscribers/destroy.json";
     pub const ADD_LIST: &'static str = "https://api.twitter.com/1.1/lists/members/create_all.json";
+    pub const REMOVE_LIST: &'static str = "https://api.twitter.com/1.1/lists/members/destroy_all.json";
 }
 
 pub mod place {