+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   - New function `subscribe` to subscribe to a list
   - New function `unsubscribe` to unsubscribe from a list
   - New function `add_member_list` to add multiple users to a list
+  - New function `remove_memebr_list` to remove multiple users from a list
 - New enum `service::ListMethod` and `list` field in `service::RateLimitStatus` to contain the rate
   limit info for these methods